Split keyboard module into several submodules.
