+++ crates/bevy_input/src/core/mod.rs
@@ -1,6 +1,7 @@
 //! Core input functionality module
 
 pub mod axis;
+pub mod element_state;
 pub mod input;
 pub mod input_plugin;
 pub mod system;
@@ -12,6 +13,7 @@ pub use prelude::*;
 pub mod prelude {
     pub use super::{
         axis::Axis,
+        element_state::ElementState,
         input::Input,
         input_plugin::InputPlugin,
         system::{exit_on_esc_system, ExitOnEscapeState},
