+++ crates/bevy_input/src/mouse/mod.rs
@@ -1,5 +1,4 @@
-use super::keyboard::ElementState;
-use crate::core::Input;
+use crate::core::{ElementState, Input};
 use bevy_app::prelude::{EventReader, Events};
 use bevy_ecs::{Local, Res, ResMut};
 use bevy_math::Vec2;
